+++ lib/new_relic/transaction/complete.ex
@@ -140,6 +140,7 @@ defmodule NewRelic.Transaction.Complete do
       |> Map.merge(NewRelic.Config.automatic_attributes())
       |> Map.put(:"nr.apdexPerfZone", Util.Apdex.label(apdex))
       |> Map.put(:total_time_s, tx_attrs.duration_s + concurrent_process_time_ms / 1000)
+      |> Map.put(:process_spawns, length(process_spawns))
 
     {[segment_tree], tx_attrs, tx_error, span_events, apdex, tx_metrics}
   end